chore: move form validation to search_controller
